+++ CHANGELOG.md
@@ -13,6 +13,7 @@ and [Conventional Commits](https://www.conventionalcommits.org/en/v1.0.0/).
 
 ### Fixed
 - [docs] Fix Rustdoc lints and warnings
+- [remote] Restart client on user token expiration
 
 ## [0.5.0] - 2024-12-09